pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "funding-service-design-utils"
version = "1.0.14"
authors = [
  { name="DLUHC", email="FundingServiceDesignTeam@levellingup.gov.uk" },
]
description = "Utilities used by the DLUHC Funding Service Design Team"
readme = "README.md"
license = { file="LICENSE" }
requires-python = ">=3.10,<4.0"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "Flask-Babel>=2.0.0,<3.0.0",
    "PyYAML>=6.0,<7.0",
    "python-dotenv>=0.20.0,<0.21.0",
    "rich>=12.4.4,<13.0.0",
    "Flask>=2.1.1,<3.0.0",
    "python-json-logger>=2.0.2,<3.0.0",
    "gunicorn>=20.1.0,<21.0.0",
    "pytz>=2022.1",
    "PyJWT[crypto]>=2.4.0",
    "sentry-sdk[flask]>=1.9.9,<2.0.0"
]

[project.urls]
"Homepage" = "https://github.com/communitiesuk/funding-service-design-utils"
